+++ Sources/KeyboardKit/_Deprecated/KeyboardLocale+Deprecated.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public extension KeyboardLocale {
+
+    @available(*, deprecated, message: "Use locale.localizedName instead")
+    var localizedName: String {
+        locale.localizedName
+    }
+}